cards section is implemented and some updates are made

+++ src/main/java/com/flexbank/ws/entity/LoanNotification.java
@@ -33,7 +33,7 @@ public class LoanNotification {
     private Integer id;
 
     @Column(name = "amount")
-    private int amount;
+    private Double amount;
 
     @Column(name = "due_to")
     private LocalDate dueTo;
